Properly handle focus when a pane is showing/hiding/levitating
The default focus systems, no matter the normal app focus or the a11y
focus, don't work well with the pane scaffold we created as we will
dynamically showing/hiding/switching/levitating panes and we need to
specify certain semantics so the focus systems can be aware of the
intention or the expected behavior when those pane operations happen.

To achieve that goal, this CL:

1. Explicitly requests focus when the current destination pane changes.
2. Block underlaying panes from being focused when a pane is levitating
   with a scrim.

Note that this also solves multiple a11y issues we have regarding the
a11y focus shifting during pane operations.

Relnote: "Adds an indicator to denote a pane should be interactable or
  not. Also uses this indicator to disable accessibility access for
  underlying panes when a levitated pane is shown with a scrim."

Test: to add in a separate CL
Bug: 437912187
Bug: 323387770
Bug: 437904302
Bug: 437901930
Change-Id: If36f3851addfe8ba48386440641c2462ee866a52