Add extensive logging to image capture and increase effect test timeout
Image capture failures were difficult to debug due to a lack of detailed
logging throughout the pipeline. It was challenging to trace a single
request's lifecycle or identify performance bottlenecks. Additionally, a
related test in EffectsFragmentDeviceTest was prone to timing out on
some devices in rare cases.

This change introduces comprehensive logging and tracing to improve the
debuggability of the image capture process and enhances test stability.

Specifically, this commit:
*   Introduces a `CameraXTracer` utility to wrap time-sensitive code
    paths in trace events, prefixed with 'CX:' for easy filtering, as
    per the current norm of CameraX traces.
*   Adds detailed logs with request IDs at key stages within
    `ProcessingRequest`, `CaptureNode`, `ProcessingNode`, and
    `FileUtil`.
*   Increases the capture timeout in `EffectsFragmentDeviceTest` to
    reduce flakiness and improve test reliability.

Bug: b/433648873
Test: EffectsFragmentDeviceTest, ImageCaptureTest
Change-Id: Ib92115e8ba18d21c04636830facb63a66ec24739