Relnote: Refactor PdfPageContent to support generic selections
This change generalizes the page content selection mechanism in the PDF viewer. The existing implementation was tightly coupled with text selection.

To provide a more flexible and extensible foundation, this CL refactors `PdfPageContent` and `PageSelection` to be more abstract. This will allow them to handle various types of selections (e.g., text, links) and can be extended in the future to support other content like images.

Bug: 437845185
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I28f167f9501ff9b78e883794d0e33fc9ac26a655