Add overloads to codec to support nullable types
Add overloaded encodeToSavedState/decodeFromSavedState functions to support nullable types. We already support nullable types for nested properties, this change is to support nullable types for root objects as well. The representation for `null` at root will be:

```
{""=null}
```

Note that `null` is considered a primitive in our codec so `ClassDiscriminatorMode.ALL_OBJECTS` has no effect on `null` values. For non-null values we record the runtime non-null type even when the static type is nullable, e.g. when encoding `Person("Tom")` with static type `Person?` the "type" we record is still "Person".

Bug: 439527454
Relnote: Add overloads to codec to support nullable types
Test: SavedStateCodecTest
Test: SavedStateCodecClassDiscriminatorAllObjectsTest
Test: SavedStateCodecClassDiscriminatorPolymorphicTest
Test: SavedStateCodecEncodeDefaultsTest
Change-Id: I7906231207e2b25c64111123c0c9f22802c87139