Stop creating stack trace while assigning ProtectionGroup
The logic was added for debugging. The root cause has been found that
robolectric didn't make the scheduled message to be executed after the
current one, but executed the incoming message directly, and caused the
calling sequence unexpected.

This CL removes the stack trace for better performance, and moves
removeProtectionViews closer to the creation of ProtectionGroup to
make the logic more robust.

Fix: 437035243
Test: ProtectionLayoutActivity
Change-Id: Ide98854e2aaae878cc52e891a11378eb2ae2d207