Prevent test Activity from recreating on rotation
EmptyTestActivity were recreated when the device was
rotated during tests.

This occurs because a configuration change (like rotation)
causes the Activity to be destroyed and recreated by default.
When the Activity is destroyed, its window and any associated
state, such as wake locks are lost.

This change prevents this recreation by adding
'android:configChanges="orientation|screenSize"' to the
test activity's manifest entry.

Bug: 423536869
Test: ZslDeviceTest
Change-Id: I762bbd10bd4b660c0bbaab65d0b00b18cf2602b0