Update Subspace Composables to have SubspaceModifier for sizing.
Constraints parameter is replaced with SubspaceModifier + size-related modifiers to follow Compose convention. In XR, the space is unbounded but it does not mean that it is an ideal size to design XR content. Unless allowUnboundedSubspace is set true, the recommendedContentBoxSizeInFullSpace is applied to the ApplicationSubspace, Subspace and GravityAlignedSubspace using the internal recommendedSizeIfUnbounded modifier.

Test: added unit tests and refactored existing impacted tests.
Bug: 433331675
RelNote: Introduce SubspaceModifier to Subspace composables and replace constraints parameter with SubspaceModifier with size-related SubspaceModifiers. If allowallowUnboundedSubspace is true, Subspaces can still have unbounded constraints.

Change-Id: Ib06e6d268a161e8e132766668df5a48c9659bd6b