Codify retention behavior for key parameters
Any `key` inputs given to `retain` will always stay in memory for as
long as the retained value does to ensure that they may be accurately
compared when restoring state. The same care has to be given to key
inputs as retained values when avoiding memory leaks. Retained keys do
not get callbacks to RetainObserver (mirroring the behavior of remember
and RememberObserver).

This CL adds documentation about the retention behavior of keys to
`retain` and expands lint inspections to flag un-retainable types used
as key inputs.

Test: RetainDetectorTest
Bug: b/177562901
Relnote: "Clarified in documentation and in tooling that keys passed to
retain are also retained. Avoid passing keys to retain that will lead
to a memory leak."

Change-Id: Ib553beecc3f574424d5ce9d7cc66d48021d87558