Restrict drag handle support to text selections.
The current implementation of selection drag handles only supports pure
text selections. Enabling them for other types, such as mixed-content
selections, is not fully supported and can cause unexpected behavior.

To ensure stability, this change disables drag handles for any selection
that is not exclusively a `TextSelection`. This functionality will be
extended to support other selection types in a future change.

Bug: 441018502
Test: ./gradlew :pdf:integration-tests:connectedAndroidTest
Change-Id: I1ecf6725e587a27c2a24100962c3f1c2f45b027f