Align ScrollIndicator direction with content layout direction by default.
Currently, when a developer uses a `ScalingLazyColumn` or `LazyColumn` with `reverseLayout = true`, the `ScrollIndicator` is not reversed by default. This leads to a confusing user experience where the scrollbar moves in the opposite direction of the content.

To fix this, the developer must manually pass `reverseDirection = true` to the `ScrollIndicator`. This is error-prone and not an intuitive default.

This change aligns the default behavior by making the `reverseDirection` parameter on the `ScrollIndicator` composable default to the value of `state.layoutInfo.reverseLayout` for `ScalingLazyListState` and `LazyListState`.

Test: Added screenshot tests.
Bug: 441489028
Relnote: "Align ScrollIndicator direction with content layout direction by default."
Change-Id: I0da0f38228025c59da02f9cc700773085c96c2f4