Store saved resources on disk when ProtoLayoutScope is used on older renderers
For older renderers, when resources are not sent with Tile,
we save resources in TileService in memory. If
onTileResourcesRequest isn't called before service is about to
be destroyed, we would loose them. To bridge this, we are going
to save resource on disk too, only in cases when scope is used on
older renderer. This is acceptable disk operations, as most of
tiles are not multi-instance, and we will only do it for older
renderers (as new renderers will bundle Resources in Tile
response). We will save them in SharedPreferences so they can be
retrieved once we need to read them in onTileResourcesRequest.

If for some reason saving fails, users won't see the error, as
onTileResourcesRequest wouldn't have anything, so provider will
call service's method.

Fixes: 428692502
Test: Added
Relnote: "Save resources used from ProtoLayoutScope to be correctly sent in onTileResourcesRequest for older renderers that don't bundle it within Tile response."
Change-Id: I063a804ab158869a03d5596ab7da5fc0b32da920