Do not mark nodes as placed during alignment lines calculation.
We were setting isPlaced and calling callbacks with incorrect coordinates during the alignment calculation placement pass. It is wrong, as the node should only be considered placed when placeable.place() is called for this node and all of its parents.

It is a third attempt to land this fix, which is causing incorrect coordinates to be reported by our coordinates tracking mechanism. Previous attempts to merge it caused b/436622681 and b/441974353. Both issues are fixed and covered with tests. Also, this time I am trying to split the change more:
* The fix to only call updates on RectManager when the node is placed is done separately in aosp/3758268. Together with this cl, the alignment lines related behavior is fully fixed for onLayoutRectChanged callbacks, and extra tests are added
* This time I am not trying to fully fix onPlaced dispatching logic, as it should be properly reconsidered separately. This change still fixes some of the onPlaced use cases, and I've filed b/442545780 for the rest of them.
* This change fixes all the found bugs for the use cases when the coordinates used in the placement scope.

Test: PlacedChildTest, OnGlobalRectChangedTest, PlacementLayoutCoordinatesTest

Bug: 434151773
Bug: 433494436
Change-Id: Ia4c6b570e05e8e0929f7fc49238e186150cea5aa