Remove the drag-to-resize feature
Due to a11y review and API coucil feedback, we decided that the API
surface of the feature is not mature enough to be released. Removes the
relevant APIs from public API surface and we will re-check them into
the upcoming 1.3 alpha versions instead.

Relnote: Remove the drag-to-resize feature from the public API surface
Test: compilation
Bug: 437953743
Bug: 442636084
Change-Id: Ic85ba48ddc8f358abe1996c0d4c8a5c2b938f380