Use sequentialExecutor for CameraCallbackMap
Under camera-camera2, callbacks are fired with a SequentialExecutor.
This has a side effect of catching any RuntimeExceptions if they occur
during the callback.

CameraPipe Integration uses the backgroundExecutor, which comes from the
camera executor. This directly contrasts camera-camera2 which creates a
SequentialExecutor wrapping the camera executor.

By using the sequentialExecutor here, the callback executor is now
aligned with camera-camera2.

Bug: 416330722
Test: Run affected tests
Change-Id: If53d2a109c86719e2067616032b31c01a9400884