Ignore suspending function type replacement if FunctionN type is not found.
Variance resolution and suspend function type replacement is only needed for Java code generation on JVM targets and fail for non-JVM as the family of kotlin.jvm.function.* interfaces are not available, unfortunately since both Java and Kotlin names are created regardless of the code generation language, we need to ignore the missing type and instead perform resolution on the kotlin.coroutines.SuspendFunctionN even though such name should not be used for Java code generation.

Overall this is a workaround for XProcessing not being configurable to do processing based on the target generating language (b/254135327).

Bug: 442220723
Test: KMP QueryTest on native
Change-Id: I3bcc2db49dfcfc84388c64beafdb3961bff160d8