Add onInfoChanged callback to NavigationEventInput
Introduces a new `onInfoChanged` callback to the NavigationEventInput
API. This callback is invoked whenever the NavigationEventInfo history
changes, providing the input with the current, back, and forward
contextual information.

RelNote: "Add `onInfoChanged` callback to `NavigationEventInput` to
notify listeners of changes to the navigation history. This provides
the full context of the current, back, and forward stacks, enabling
Inputs to react to the contextual information."

Fixes: 443282983
Test: NavigationEventDispatcherTest#onInfoChanged_onStateChange_notifiesInput
Change-Id: I69a8bf349e514983b821ab9905d66ac5d9889ac9