Link KeyboardOption type to autofill ContentType
Sets the content type in semantics for BasicTextField and CoreTextField based on the keyboard type, i.e. when the keyboard mode is set to email, this will set the autofill content hint to be email address under the hood for the developer.

This allows autofill services to correctly identify the field type. See go/extended-compose-autofill  for more details.

Test: added in BasicTextFieldSemanticsTest.kt and PerformAndroidAutofillManagerTest.kt. Also manually verified that existing autofill samples still work.
Change-Id: I85a524c62ddaa36b208a9c59a32ace7c3ff2255c