Fix issue with kotlin metadata for companion object members.
This CL fixes https://issuetracker.google.com/443344468.

This CL really fixes a few issues that are both related
to missing kotlin metadata for methods that originate in
the companion object.

First, `XMethodElement.isExtensionFunction()` is incorrect for
extension functions in companion object. Second, `XMethodElement.isPropertyFunction()` is incorrect
for property getter and setter functions corresponding to
properties in a companion object. Both of these issues are
due to missing kotlin metadata for methods that come from
companion object.

In addition, there's a similar issue for `XFieldElement` for
fields that come from a companion object.

This CL fixes both issues by checking the companion object
for kotlin metadata if its not found in the enclosing element.

Bug: https://issuetracker.google.com/443344468
Test: XExecutableElementTest#extensionFun()
Change-Id: I3d7e95feef674273a7575983d0b175bd5af45f81