Update rotate assertions to use tolerance and add missing unit tests.
This will guard the current rotate modifier behaviors e.g. the rotate modifier combines rotations in `local * parent` order and not vice versa.

Bug: 444111572
Test: new tests added and passed.
RelNote: assertRotationInRootIsApproximatelyEqualTo asserts that the layout of this node has a rotation in the root composable that is approximately equal to the given rotation within a small tolerance.
Change-Id: Ieec32534510714a11f5a91e4715415ae9345bca0