Introduce UnplacedStateAwareModifierNode
This modifier type was already introduced as part of aosp/3399519 internally in order to unblock the use case from JB. However, we wanted for more use cases to arise before we decide on the api shape for it. As we just used this modifier for Modifier.onVisibilityChanged(), it is now clear there is a valid use case for it, and without this public api users can't know when the node is not placed anymore, while we provide a few different apis to react on a node being placed.
Another discussion was about this being a separate interface or just a new fun in LayoutAwareModifierNode, and now we are leaning towards it being a separate interface. Few reasons for it:
* When we combine many callbacks in one modifier type, we can't really know which one of them are important to exact implementation, so we have to trigger all the callbacks, even for the ones which has no-op implementation. It forces us to do extra work.
* It is also hard to handle the modifier autoInvalidation in the most efficient way, as we don't know what callbacks are important for the exact implementation. For example, when we add a new modifier, which is reacting on unplaced event we don't need to remeasure, but when the modifier needs onRemeasured callback, we need to.
* In future we might even consider completely deprecating LayoutAwareModifierNode if we decided that we don't really need to recommend onPlaced callback for our users (the whole story is confusing here as onGloballyPositioned and onPlaced have the same api shape and similar contract, why do we have both?)

Aside from renaming and making this interface public I've also fixed the case when the callback wasn't called when the placed node is being removed from the tree.

Bug: 309776096
Test: UnplacedStateAwareModifierNodeTest
Relnote: UnplacedStateAwareModifierNode was introduced. It provides a callback called when the previously placed layout is not placed anymore.
Change-Id: I8fdd82ad7b8b750209b07ba4b54af4152910be8e