Handle camera open hangs
Some platforms unfortunately can have CameraManager.openCamera() that
blocks indefinitely. When this happens, this blocks all subsequent
camera operations, including the camera shutdown process.

Refactor CameraStateOpener to handle this with a better, more robust
mechanism. We compartmentalize the procedure into 3 jobs, and promptly
time out the openCamera() call when it takes more than 3s to return. We
also add a dedicated error code for that.

Bug: 433662146
Test: Blocking various places
Change-Id: Ib90d647c3e371d7f5e7eecd6483194cd9bd587ce