[acms-fixit] Fix memory leak and observer lifecycle in PreviewView
Ensures to trigger the onSurfaceNotInUse callback for the previous SurfaceRequest if it is still not serviced. This will make the PreviewView do the clean flow properly to avoid the memory leakage problem.

Also, update the LiveDataObservable to make it correctly forwarding the results to all the existing observers.

Relnote: "Fixed a memory leak in PreviewView where it could prevent its Activity from being garbage collected. This happened if a new SurfaceRequest arrived before the previous one is handled."

Bug: 443112512
Test: PreviewViewTest and tests in viewtestapp
Change-Id: I69622e1ce802c943f59640219e706d7cd9f06b06