Add a null check on RawQuery parameter
For better interoperability of Kotlin codegen from Java sources, add a null-check to the raw query parameter if it is non-null. Even though the processor will raise an error if it is explicitly null, it can have unknown nullability due to Java.

Also move the missing query param code gen logic from the compatibility path to the driver API side.

Bug: 339809512
Bug: 438041176
Test: DaoKotlinCodeGenTest
Change-Id: I72822d53e0bdf78a048c678d76ad87c313e0f5c2