Only override routes when the new route supports newer clients
Media Route providers publish MediaRouteDescriptors as a list,
which the MediaRoute2ProviderServiceAdapter converts to a map
using the descriptor ID as the key. Previously, if the list
contained duplicate IDs, the descriptor that appeared last in
the list would always override any preceding entries in the map.
This implementation ignored client compatibility.

This change updates the logic to check the supported client
version before overriding. The existing descriptor will only be
replaced if the new descriptor supports a more recent client
version range.

Change-Id: Id712d3aa76d207ce4e7e1c3cdb61345046c410b8
Bug: 444697229
Test: presubmit and demo app