Add SceneState to Navigation Runtime
Moving most of scene logic from the NavDisplay into a separate
SceneState object that finds the needs properties via a
rememberSceneState call, and then can be passed into the NavDisplay
separately.

This also rename NavDisplayInfo into SceneInfo and moves it to runtime.

This moves SinglePaneScene into runtime as well.

RelNote: "Added a new SceneState object to help manage scenes. This also
provides a new NavDisplay overload that takes the SceneState and the
NavigationEventState."
Test: Added SceneStateTest
Bug: 444479133

Change-Id: Idfb4648b798e71000277448e355a00e0fb1e4fd1