Use AnimatedContentTransitionScope<Scene<T>>
Update NavDisplay so that instead of using
a Pair<KClass<Scene<T>>, Any> for the
AnimatedContentTransitionScope, simply
use the Scene itself.

This also fixes how the
LocalEntriesToRenderInCurrentScene is
calculated to strictly match the z-order
of scenes and also handle overlay scenes.

Test: existing tests pass
Relnote: "`NavDisplay` now allows you to
customize the transitions based on what
`Scene` you are going to and from by looking
at the `Transition`'s `currentState` and
`targetState`."
BUG: 443872322

Change-Id: I906cc1a8305e1d6cabfdfe8a4ac555e330863fbf