Implement JxrRuntime lifecycle methods for Spatial Runtimes
This change implements the `JxrRuntime` lifecycle methods in `SpatialSceneRuntime` and `SpatialRenderingRuntime`.

The startRendering and stopRendering are merged in the lifecycle.

And also add `dispose()` method to the `SpatialEnvironmentFeature` interface and refine dispose and destroy flow in `SpatialRenderingRuntime`, ``SpatialEnvironmentImpl` and `SpatialEnvironmentFeatureImpl` to ensure proper resource cleanup.

Tests are added to verify the new lifecycle method implementations and the disposal logic for `SpatialEnvironmentFeature`.

Test: unit test, scenecore testapp
Change-Id: I3d0b51509c45de64f03d7b664012b817d3de3b08