Optimisation - Reducing rendering time of first page after bitmap has been loaded
This Cl refactors the setting empty highlights,
An invalidation signal was being triggered by the hot flow that sets highlights, even when the highlights were empty.
This caused an extra, redundant onDraw pass every time a PDF was loaded.
To fix this, the code now checks if the highlights are empty before triggering the invalidation flow.

Test: Build and run on local device
Bug: 445598135
Change-Id: Ibb30a6036659211878b9596d5282ad38aae3288c