Make NavigationEventTransitionState.Idle an object
Change NavigationEventTransitionState.Idle from a class to a
Kotlin object.

This ensures it acts as a singleton, as there is no need for
multiple instances of a stateless 'Idle' value. All call sites
are updated to reference the 'Idle' object directly instead of
instantiating 'Idle()', preventing unnecessary allocations.

RelNote: "Use the `NavigationEventTransitionState.Idle` singleton
object directly instead of instantiating `Idle()`."

Bug: 444734264
Test: NavigationEventDispatcherTest.kt
Change-Id: Ic7d9ec5eb9a1c426cd78342a9244df324f3147bd