Opt for EmojiCompatInitializer from androidx.emoji2
The `EmojiCompatInitializer` was causing a runtime class error when its visibility was `internal`. The class is referenced in the AndroidManifest.xml and thus is expected to be accessible to code external to this package. Additionally it uses emoji2-bundle which adds 10+MB to the build artifact.

This removes the bespoke implementation of `EmojiCompatInitializer`, its reference in AndroidManifest.xml and its dependence on emoji2:emoji2-bundle, replaced with emoji2:emoji2. It is not necessary to use a provider xml attribute as described in the docs for EmojiCompatInitializer:

> This is enabled by default by including the :emoji2:emoji2 gradle artifact.

The crash is not observed in the unmodified wear integration demo, but a similar stacktrace can be observed by changing the visibility to private, at which point the application has a runtime error upon startup (NOT a compile-time error!)

Bug: 444422736
Relnote: Compat rendering of emoji in warped text when SDK >= 29 and SDK < 34
Test: ./gradlew :wear:compose:integration:demos:installDebug && adb shell am start androidx.wear.compose.integration.demos/.DemoActivity
Change-Id: I589b4a9427b93989fde0bcd937fe5059cbc9a011