Add shortcut driver APIs for transactions
The extension functions withReadTransaction() and withWriteTransaction() are useful for operation that are to be completely made within a transaction. The two functions use the recommended transaction types when reading and writing.

This change also makes Room's OptIn ExperimentalRoomApi common, since it was Android only.

Also migrates InvalidationTrackerFlowTest to use driver APIs.

Bug: 438041176
Test: InvalidationTrackerFlowTest
Relnote: Add shortcut functions withReadTransaction() and withWriteTransaction() for starting a read or write transactions with driver APIs.
Change-Id: Ief0e5d2004701424774abc7e74a4fd945426caeb