Convert VerticalStack tests to use indirect touch input.
This CL also moves IndirectTouchHelper outside of the `list` package, so
that it can be shared across the Glimmer subpackages.

At the moment, focus has to be requested after every indirect touch
swipe, so that the current item at the top of the stack gets focus. Once
stack supports moving focus automatically, the test code can be
simplified.

Bug: 413429531
Test: VerticalStackTest, StackStateTest
Change-Id: I359656afccfd1061bc14b21f36763c1cff7b3c15